+++ index.cjs
@@ -1,0 +1,2 @@
+const Figurecon = require('./src/index.cjs');
+module.exports = Figurecon;